modelo y controller modificado en tipo de dato

+++ Spring Boot Project/src/main/java/com/ejemplo/tiendaalamano/model/Categorias.java
@@ -21,7 +21,7 @@ public class Categorias implements Serializable{
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id_categoria;
+	private Long id_categoria;
 	
 	@Column(length = 15, nullable=false)
 	private String nombre_categoria;
@@ -31,11 +31,11 @@ public class Categorias implements Serializable{
 	private Subcategorias subcategorias;
 	
 	
-	public int getId_categoria() {
+	public Long getId_categoria() {
 		return id_categoria;
 	}
 
-	public void setId_categoria(int id_categoria) {
+	public void setId_categoria(Long id_categoria) {
 		this.id_categoria = id_categoria;
 	}
